Moved id_discord to BIGINT(20)

+++ models/utilisateur.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Utilisateur = sequelize.define('Utilisateur', {
     prenom: DataTypes.STRING,
     nom: DataTypes.STRING,
-    id_discord: DataTypes.INTEGER
+    id_discord: DataTypes.BIGINT(20)
   }, {});
   Utilisateur.associate = function(models) {
     Utilisateur.belongsTo(models.Role);
